remove whoosh dependency and replace it whith miniwhoosh
Uploading documents from the frontend is now also possible using the
miniwhoosh endpoints. For this the document needs to be processed in the
frontend.
Searching for a document using the miniwhoosh library is now done via
the endpoint api/miniwhoosh/search. Note that if the miniwhoosh index is
encrypted, then the query sent to the endpoint also needs to be
encrypted with the same key material used when encrypting the index.